+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Both client libraries are pre-1.0, and they have separate versioning.
 ## Unreleased
 
 - Added `env` parameters to several methods, as a convenience for passing environment variables into Sandboxes, etc.
+- Added `Sandbox.getTags()` (JS) and `Sandbox.GetTags()` (Go).
 
 ## modal-js/v0.3.23, modal-go/v0.0.23